basic stats page (not linked) for me to check on daily data importer job results
