+++ lib/analytics.js
@@ -1,24 +1,30 @@
 import { fetchGraphQL } from './graphql';
 
-const HASURA_GET_DATA_IMPORTS = `query FrontendGetDataImports {
-  ga_data_imports(order_by: {created_at: desc}) {
-    created_at
+const HASURA_GET_DATA_IMPORTS = `query FrontendDataImports($end_date: date, $start_date: date) {
+  ga_data_imports(where: {end_date: {_eq: $end_date}, start_date: {_eq: $start_date}}, order_by: {organization: {slug: asc}, table_name: asc}) {
     end_date
-    id
     notes
+    row_count
     start_date
-    organization_id
+    success
     table_name
-    updated_at
+    organization {
+      name
+      slug
+    }
   }
 }`;
 
 export function hasuraGetDataImports(params) {
   return fetchGraphQL({
     url: params['url'],
-    orgSlug: params['orgSlug'],
+    adminSecret: params['apiToken'],
     query: HASURA_GET_DATA_IMPORTS,
-    name: 'FrontendGetDataImports',
+    name: 'FrontendDataImports',
+    variables: {
+      start_date: params['start_date'],
+      end_date: params['end_date'],
+    },
   });
 }
